sales_pipeline/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sales_pipeline"
version = "0.1.0"
description = "Pipeline de vente Bronze-Silver-Gold"
readme = "README.md"
requires-python = ">=3.9"


dependencies = [
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pyspark" 
]